added in teleportation and grab interactable for a drawer
Added in teleportation by following the below tutorial - https://www.youtube.com/watch?v=8xIxXkXn26M&t=353s&ab_channel=FistFullofShrimp

The key steps where Adding a teleportation provider, from here added in a teleport interactor on the right hand. Then adding in the teleport anchor and adding and adjusting the xr interaction group.

For the drawer I again followed a tutorial online for this - https://www.youtube.com/watch?v=bYS35_hC6B0&ab_channel=Valem

For this the main steps where adding in a xr grab interactable and setting the movement type to velocity and also importantly setting the drawer handler as the collider in the xr grab interactable as this will be the only part which can be grabbed. Also importantly adding a layer called grab ignore ray so that the door handle can only be picked up by the hand and not the ray interactor and within the physics project settings disabling for the new layer. One of the last steps was adding in the hinge joint so that door will swing open and stay attached to the rest of the drawer